move inject_data_binmsg into rpc.hpp
