+++ rpc.hpp
@@ -8,6 +8,7 @@
 
 #include <ch_frb_io.hpp>
 #include <assembled_chunk_msgpack.hpp>
+#include <msgpack_binary_vector.hpp>
 #include <rf_pipelines_inventory.hpp>
 
 /*
@@ -232,7 +233,25 @@ public:
     MSGPACK_DEFINE(start, period, packets);
 };
 
-#include "inject_data_msgpack.hpp"
+/* This defines the 'wire protocol' for the L1 data-injection RPC call.
+ */
+struct inject_data_binmsg {
+    int beam;
+    // mode == 0: ADD
+    int mode;
+    // offset for FPGAcounts values in fpga_offset array.
+    uint64_t fpga0;
+    // should be "nfreq" in length
+    ch_frb_io::msgpack_binary_vector<int32_t> sample_offset;
+    // should be "nfreq" in length
+    ch_frb_io::msgpack_binary_vector<uint16_t> ndata;
+    // should have length = sum(ndata)
+    ch_frb_io::msgpack_binary_vector<float> data;
+    
+    MSGPACK_DEFINE(beam, mode, fpga0, sample_offset, ndata, data);
+
+    void swap(rf_pipelines::inject_data& dest);
+};
 
 
 #endif
